ai: KAIK: buffer overflow preventions * replaced sprintfs with SNPRINTFs * made some class members private

+++ AI/Skirmish/KAIK/SpotFinder.cpp
@@ -133,7 +133,7 @@ CachePoint* CSpotFinder::GetBestCachePoint(int x, int y)
 	int cacheX = x;		// / CACHEFACTOR;
 	int cacheIndex = cacheY * MapWidth / CACHEFACTOR + cacheX;
 
-	if (cacheIndex >= 0 && cacheIndex < cachePoints.size()) {
+	if (cacheIndex >= 0 && (unsigned int)cacheIndex < cachePoints.size()) {
 		if (!cachePoints[cacheIndex].isValid) {
 			MakeCachePoints();
 		}
